Split login function into crawler
