+++ routes/router.go
@@ -11,6 +11,6 @@ RegisterRoutes is a place to register rotes
 */
 func RegisterRoutes(router *gin.Engine) {
 
-	router.POST("/login", login.HandleLogin)
+	router.POST("/login", login.Controller)
 
 }
